Separate repackaging functions to common file
